+++ srbuild/project/target.py
@@ -3,6 +3,7 @@ from typing import NewType, Dict
 
 # A Dict[str, LinkedNode] that maps profile names to the LinkedNodes for a target
 class ProjectTarget(dict):
-    def __init__(self, name, *args, **kwargs):
+    def __init__(self, name, is_lib=False, *args, **kwargs):
         super().__init__(self, *args, **kwargs)
         self.name = name
+        self.is_lib = is_lib